# backend/src/routers/reports.py
# backend/src/routers/reports.py

import logging
import os
import io
from datetime import datetime
from typing import Optional
from uuid import uuid4

from fastapi import APIRouter, Depends, HTTPException, status, Query
from fastapi.responses import FileResponse

import sys
import os as os_module
sys.path.insert(0, os_module.path.dirname(os_module.path.dirname(os_module.path.dirname(os_module.path.abspath(__file__)))))

from sqlalchemy import text

from src.schemas import UserSchema
from src.routers.auth import get_current_active_user

logger = logging.getLogger(__name__)

router = APIRouter(prefix="/reports", tags=["reports"])


def generate_report_id() -> str:
    year = datetime.utcnow().year
    unique_part = str(uuid4())[:8].upper()
    return f"REP-{year}-{unique_part}"


@router.post("/generate")
async def generate_report(
    report_format: str = Query(..., pattern="^(PDF|HTML|CSV)$", description="Formato del reporte"),
    include_kpis: bool = Query(True, description="Incluir KPIs en el reporte"),
    include_alerts: bool = Query(False, description="Incluir alertas en el reporte"),
    include_returns: bool = Query(True, description="Incluir devoluciones en el reporte"),
    date_from: Optional[str] = Query(None, description="Fecha inicio (YYYY-MM-DD)"),
    date_to: Optional[str] = Query(None, description="Fecha fin (YYYY-MM-DD)"),
    channel: Optional[str] = Query(None, pattern="^(B2C|B2B)$", description="Canal de retorno"),
    current_user: UserSchema = Depends(get_current_active_user),
):
    if current_user.role not in ["SUPERVISOR", "KAM", "ADMIN"]:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="No tiene permisos para generar reportes"
        )

    from src.db.database import SessionLocal
    db = SessionLocal()

    try:
        report_id = generate_report_id()

        filters = {
            "date_from": date_from,
            "date_to": date_to,
            "channel": channel,
        }

        insert_query = text("""
            INSERT INTO reports (
                report_id, format, status, filters,
                include_kpis, include_alerts, include_returns,
                requested_by, created_at
            ) VALUES (
                :report_id, :format, :status, :filters,
                :include_kpis, :include_alerts, :include_returns,
                :requested_by, :created_at
            )
        """)

        db.execute(insert_query, {
            "report_id": report_id,
            "format": report_format,
            "status": "PROCESSING",
            "filters": str(filters),
            "include_kpis": include_kpis,
            "include_alerts": include_alerts,
            "include_returns": include_returns,
            "requested_by": current_user.username,
            "created_at": datetime.utcnow(),
        })

        db.commit()

        logger.info(f"Report {report_id} generated by user {current_user.username}")

        return {
            "report_id": report_id,
            "format": report_format,
            "status": "PROCESSING",
            "filters": filters,
            "include_kpis": include_kpis,
            "include_alerts": include_alerts,
            "include_returns": include_returns,
            "requested_by": current_user.username,
            "created_at": datetime.utcnow().isoformat(),
            "message": "Reporte generado exitosamente. Use /reports/{report_id}/download para descargarlo.",
        }
    except HTTPException:
        db.rollback()
        raise
    except Exception as e:
        db.rollback()
        logger.error(f"Error generating report: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Error al generar reporte: {str(e)}"
        )
    finally:
        db.close()


@router.get("/{report_id}/download")
async def download_report(report_id: str):
    """
    Download a generated report by its ID.
    Returns the report file as a downloadable attachment.
    """
    try:
        from src.db.database import SessionLocal

        db = SessionLocal()
        try:
            query = text("""
                SELECT report_id, format, status, file_path, filters,
                       include_kpis, include_alerts, include_returns,
                       requested_by, error_message, created_at, completed_at
                FROM reports
                WHERE report_id = :report_id
            """)

            result = db.execute(query, {"report_id": report_id}).fetchone()

            if not result:
                raise HTTPException(
                    status_code=status.HTTP_404_NOT_FOUND,
                    detail=f"Report with ID {report_id} not found"
                )

            report = {
                "report_id": result[0],
                "format": result[1],
                "status": result[2],
                "file_path": result[3],
                "filters": result[4],
                "include_kpis": result[5],
                "include_alerts": result[6],
                "include_returns": result[7],
                "requested_by": result[8],
                "error_message": result[9],
                "created_at": result[10],
                "completed_at": result[11],
            }

            if report["status"] != "COMPLETED":
                raise HTTPException(
                    status_code=status.HTTP_400_BAD_REQUEST,
                    detail=f"Report is not ready for download. Current status: {report['status']}"
                )

            if not report["file_path"]:
                raise HTTPException(
                    status_code=status.HTTP_404_NOT_FOUND,
                    detail="Report file not found"
                )

            file_path = report["file_path"]

            if not os.path.exists(file_path):
                logger.error(f"Report file not found at path: {file_path}")
                raise HTTPException(
                    status_code=status.HTTP_404_NOT_FOUND,
                    detail="Report file not found on server"
                )

            file_extension = os.path.splitext(file_path)[1].lower()
            media_types = {
                ".pdf": "application/pdf",
                ".html": "text/html",
                ".csv": "text/csv",
            }
            media_type = media_types.get(file_extension, "application/octet-stream")

            filename = f"report_{report_id}{file_extension}"

            return FileResponse(
                path=file_path,
                media_type=media_type,
                filename=filename,
                headers={"Content-Disposition": f"attachment; filename={filename}"}
            )

        finally:
            db.close()

    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Error downloading report {report_id}: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Error downloading report: {str(e)}"
        )


@router.get("/{report_id}")
async def get_report_status(report_id: str):
    """
    Get the status and metadata of a report.
    """
    try:
        from src.db.database import SessionLocal

        db = SessionLocal()
        try:
            query = text("""
                SELECT report_id, format, status, file_path, filters,
                       include_kpis, include_alerts, include_returns,
                       requested_by, error_message, created_at, completed_at
                FROM reports
                WHERE report_id = :report_id
            """)

            result = db.execute(query, {"report_id": report_id}).fetchone()

            if not result:
                raise HTTPException(
                    status_code=status.HTTP_404_NOT_FOUND,
                    detail=f"Report with ID {report_id} not found"
                )

            return {
                "report_id": result[0],
                "format": result[1],
                "status": result[2],
                "file_path": result[3],
                "filters": result[4],
                "include_kpis": result[5],
                "include_alerts": result[6],
                "include_returns": result[7],
                "requested_by": result[8],
                "error_message": result[9],
                "created_at": result[10].isoformat() if result[10] else None,
                "completed_at": result[11].isoformat() if result[11] else None,
            }

        finally:
            db.close()

    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Error getting report {report_id}: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Error getting report: {str(e)}"
        )


@router.get("/")
async def list_reports():
    """
    List all reports.
    """
    try:
        from src.db.database import SessionLocal

        db = SessionLocal()
        try:
            query = text("""
                SELECT report_id, format, status, requested_by, created_at, completed_at
                FROM reports
                ORDER BY created_at DESC
                LIMIT 100
            """)

            results = db.execute(query).fetchall()

            return [
                {
                    "report_id": r[0],
                    "format": r[1],
                    "status": r[2],
                    "requested_by": r[3],
                    "created_at": r[4].isoformat() if r[4] else None,
                    "completed_at": r[5].isoformat() if r[5] else None,
                }
                for r in results
            ]

        finally:
            db.close()

    except Exception as e:
        logger.error(f"Error listing reports: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Error listing reports: {str(e)}"
        )