# docs/specs/api_contract.yaml
openapi: 3.1.0
info:
  title: Derived API Contract
  version: 1.0.0
paths:
...
  /api/v1/reports/generate:
    post:
      operationId: post_reports_generate
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /api/v1/reports/{report_id}/download:
    get:
      operationId: get_reports_report_id_download
      responses:
        '200':
          description: Download report file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string
                format: binary
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Report not ready for download
          content:
            application/json:
              schema:
                type: object
...